requirements.txt
flask==2.0.1
flask-dance==5.1.0
python-dotenv==0.19.1
Werkzeug==2.0.1
authlib==1.3.2